+++ engine/src/parser/expression_utils.cpp
@@ -92,6 +92,7 @@ bool is_binary_operator(operator_type op) {
 	case operator_type::BLZ_STR_RIGHT:
 	case operator_type::BLZ_TO_DATE:
 	case operator_type::BLZ_TO_TIMESTAMP:
+	case operator_type::BLZ_STR_TRIM:
 		assert(false);
 		// Ternary operator. Should not reach here
 		// Should be evaluated in place (inside function_evaluator_transformer) and removed from the tree
@@ -229,6 +230,7 @@ cudf::type_id get_output_type(operator_type op, cudf::type_id input_left_type, c
 	case operator_type::BLZ_STR_SUBSTRING:
 	case operator_type::BLZ_STR_REPLACE:
 	case operator_type::BLZ_STR_CONCAT:
+	case operator_type::BLZ_STR_TRIM:
 		return cudf::type_id::STRING;
 	case operator_type::BLZ_TO_DATE:
 		return cudf::type_id::TIMESTAMP_DAYS;
@@ -306,8 +308,9 @@ operator_type map_to_operator_type(const std::string & operator_token) {
 		{"TO_DATE", operator_type::BLZ_TO_DATE},
 		{"TO_TIMESTAMP", operator_type::BLZ_TO_TIMESTAMP},
 		{"||", operator_type::BLZ_STR_CONCAT},
+		{"TRIM", operator_type::BLZ_STR_TRIM},
 		{"LEFT", operator_type::BLZ_STR_LEFT},
-		{"RIGHT", operator_type::BLZ_STR_RIGHT}
+		{"RIGHT", operator_type::BLZ_STR_RIGHT},
 	};
 
 	if(OPERATOR_MAP.find(operator_token) == OPERATOR_MAP.end()){
@@ -598,6 +601,10 @@ std::string replace_calcite_regex(const std::string & expression) {
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(HOUR), ", "BL_HOUR(");
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(MINUTE), ", "BL_MINUTE(");
 	StringUtil::findAndReplaceAll(ret, "EXTRACT(FLAG(SECOND), ", "BL_SECOND(");
+	// Flag options for TRIM
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(BOTH),", "TRIM(\"BOTH\",");
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(LEADING),", "TRIM(\"LEADING\",");
+	StringUtil::findAndReplaceAll(ret, "TRIM(FLAG(TRAILING),", "TRIM(\"TRAILING\",");
 
 	StringUtil::findAndReplaceAll(ret, "/INT(", "/(");
 	return ret;